requirements.txt
discord
validators
python-dotenv
asyncio
platform
colorama
validators

